# docs/swagger.yaml
definitions:
...
  dto.PublicUserDTO:
    properties:
      id:
        type: string
      image:
        type: string
      username:
        type: string
    type: object
...
  dto.UserDTO:
    properties:
      email:
        type: string
      id:
        type: string
      image:
        type: string
      username:
        type: string
    type: object
  dto.UserTokenDTO:
    properties:
      token:
        type: string
      user:
        $ref: '#/definitions/dto.UserDTO'
    type: object
  models.User:
    properties:
      createdAt:
        type: string
      email:
        type: string
      image:
        type: string
      password:
        type: string
      role:
        type: string
      updatedAt:
        type: string
      userID:
        type: string
      username:
        type: string
    type: object
  utils.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  utils.MessageResponse:
    properties:
      message:
        type: string
    type: object
info:
  contact: {}
paths:
...
  /api/v1/auth/csrf:
    get:
      description: Generates a CSRF token for the authenticated user
      responses:
        "200":
          description: CSRF token generated successfully
          schema:
            $ref: '#/definitions/utils.MessageResponse'
        "401":
          description: unauthorized
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "403":
          description: forbidden
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      summary: Generate a CSRF token
      tags:
      - Authentication
  /api/v1/auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate a user using their username and password. On success,
        returns an authentication token.
      parameters:
      - description: User login details
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.User'
      produces:
      - application/json
      responses:
        "200":
          description: Login successful with token
          schema:
            $ref: '#/definitions/dto.UserTokenDTO'
        "400":
          description: Invalid request body or missing fields
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "401":
          description: Invalid username or password
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "500":
          description: Failed to return token
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      summary: User Login
      tags:
      - Authentication
  /api/v1/auth/logout:
    post:
      description: Clears the access token cookie to log the user out.
      responses:
        "200":
          description: Logout successful
          schema:
            $ref: '#/definitions/utils.MessageResponse'
        "500":
          description: Failed to log out
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      summary: Log out user
      tags:
      - Authentication
  /api/v1/auth/register:
    post:
      consumes:
      - application/json
      description: Register a new user with a unique username, email, and password.
        On success, returns a user token.
      parameters:
      - description: User registration details
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.User'
      produces:
      - application/json
      responses:
        "200":
          description: User registration successful with token
          schema:
            $ref: '#/definitions/dto.UserTokenDTO'
        "400":
          description: Invalid request body or missing fields
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "500":
          description: Failed to return token
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      summary: Register a new user
      tags:
      - Authentication
...
  /api/v1/users/{user_id}:
    put:
      consumes:
      - application/json
      description: Update user profile information such as username and email. Requires
        a valid user ID in the request context.
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: string
      - description: Updated user details
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.User'
      produces:
      - application/json
      responses:
        "200":
          description: User update successful
          schema:
            $ref: '#/definitions/dto.UserDTO'
        "400":
          description: Invalid request body or missing fields
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "401":
          description: User not authenticated
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "403":
          description: Not enough permissions to update user details
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "404":
          description: User ID not found in context
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "500":
          description: Failed to update user details
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      summary: Update user details
      tags:
      - User
  /api/v1/users/{user_id}/image:
    post:
      consumes:
      - multipart/form-data
      description: Upload a profile image for the user. The image file should be in
        a supported image format.
      parameters:
      - description: Profile image file
        in: formData
        name: file
        required: true
        type: file
      - description: User ID
        in: path
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Image uploaded successfully
          schema:
            $ref: '#/definitions/dto.UserDTO'
        "400":
          description: Invalid file format or missing file
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "401":
          description: User not authenticated
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "403":
          description: Not enough permissions to upload image
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "500":
          description: Failed to upload image
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      summary: Upload profile imag
      tags:
      - User
  /api/v1/users/{username}:
    get:
      description: Retrieves public profile details for the specified username.
      parameters:
      - description: Username
        in: path
        name: username
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: User details retrieved successfully
          schema:
            $ref: '#/definitions/dto.PublicUserDTO'
        "400":
          description: Invalid or missing username
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      summary: Get user by username
      tags:
      - User
  /api/v1/users/me:
    get:
      description: Retrieves profile details for the currently authenticated user.
      produces:
      - application/json
      responses:
        "200":
          description: User details retrieved successfully
          schema:
            $ref: '#/definitions/dto.UserDTO'
        "401":
          description: User not authenticated
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "500":
          description: Failed to retrieve user details
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      summary: Get current user details
      tags:
      - User
swagger: "2.0"